requirements.txt
# Rice Quality Assessment Challenge - Requirements
# Deep Learning framework
torch>=2.0.0
torchvision>=0.15.0

# Vision models
timm>=0.9.0

# Image augmentation
albumentations>=1.3.0

# Data processing
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
scipy>=1.10.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Image handling
Pillow>=9.5.0

# Progress bars
tqdm>=4.65.0

# Optional: for Jupyter notebooks
# jupyter>=1.0.0
# ipywidgets>=8.0.0
